week46_wednesday/fullstack_demo/server/requirements.txt
altgraph==0.17.4
asttokens==2.4.1
blinker==1.9.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
executing==2.1.0
Flask==3.1.0
flask-cors==6.0.1
Flask-MySQLdb==2.0.0
idna==3.10
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Markdown==3.7
markdown2==2.5.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==24.1
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
requests==2.32.3
six==1.16.0
stack-data==0.6.3
tkhtmlview==0.3.1
tkinterdnd2==0.4.2
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3